pyproject.toml
[project]
name = "db-agent-mcp"
version = "1.0.0"
description = "Database Agent MCP server with support for multiple database engines"
authors = [
  { name = "Database Agent Team", email = "noreply@example.com" }
]
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
  "fastmcp==2.11.3",
  "pydantic>=2.5.0",
  "psycopg2-binary>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/db_agent"]

[project.scripts]
db-agent = "db_agent.server:main"
